v2 uwu

+++ src/app.ts
@@ -1,7 +1,7 @@
 import sqlite3 from "sqlite3"
 import { open } from "sqlite"
 import type { Database } from "sqlite"
-import { writeFile, copyFile } from "fs/promises"
+import { writeFile, copyFile, lstat } from "fs/promises"
 import { existsSync, mkdirSync } from "fs"
 import cliProgress from "cli-progress"
 import { validate } from "./fileModel"
@@ -21,7 +21,10 @@ async function starto() {
 		let file = await parseFile(res, db, "./data/files/")
 		if (file) {
 			await writeFile(`./output/jsons/${file.id}.json`, JSON.stringify(file, null, 2))
-			await copyFile(`./data/files/${file.name}`, `./output/files/${file.name}`)
+			let stat = await lstat(`./data/files/${file.name}`)
+			if (!existsSync(`./output/files/${file.id}`) || stat.size < 0) {
+				await copyFile(`./data/files/${file.name}`, `./output/files/${file.name}`)
+			}
 			bar1.increment()
 		} else {
 			bar1.setTotal(bar1.getTotal() - 1)
@@ -78,8 +81,8 @@ async function parseFile(
 			id: file.name.split(".").shift(),
 			mime: file.mime,
 			time: new Date(file.time + config.EPOCH).toISOString(),
-			selfDestruct: file.destruct === null,
-			destruct: file.destruct ? new Date(file.destruct + config.EPOCH).toISOString() : null,
+			selfDestruct: !!file.destruct,
+			destruct: !!file.destruct ? new Date(file.destruct + config.EPOCH).toISOString() : null,
 			ip: file.ip,
 			original: file.original,
 			outviews,
